Add service Makefile and .envrc scaffold
This is for each of the services in terraform-root-modules. Each service
includes this Makefile for wrapping terraform commands. The reason to
have these inside $stage.example.co is so that Atlantis can pick up
changes and know what to run a plan/apply against.

Note that as we now grab our root modules at run time, there is no need
to bundle them into the $stage.example.co geodesic container, so this
commit removes the docker multi stage build for stuffing
terraform-root-modules into the container at build time.

Also note that this means we can pull different module versions of
terraform-root-modules per service and also means we have a lot more
places to update it should we want to bump the root modules for every
service.