No multi-stage-build for terraform-root-modules
Bundling terraform-root-modules into the container at build time means
that Atlantis has no chance of knowing when a thing has changed and
therefore no idea what it needs to plan/apply. The way we get around
this is have a hook directory in conf for each terraform-root-module
with the version of the module. This means Atlantis can tell when a
version has changed and do the right thing. It takes some hookup in the
form of Makefile glue between the Makefile in `/conf` here, and a
Makefile in each terraform-root-module service. Note that we can’t use
symlinks here as Atlantis won’t follow them.