ECS Atlantis + tfenv + new goodies (#58)
* Update geodesic to 0.67.0

Mainly for Atlantis changes

* No multi-stage-build for terraform-root-modules

Bundling terraform-root-modules into the container at build time means
that Atlantis has no chance of knowing when a thing has changed and
therefore no idea what it needs to plan/apply. The way we get around
this is have a hook directory in conf for each terraform-root-module
with the version of the module. This means Atlantis can tell when a
version has changed and do the right thing. It takes some hookup in the
form of Makefile glue between the Makefile in `/conf` here, and a
Makefile in each terraform-root-module service. Note that we can’t use
symlinks here as Atlantis won’t follow them.

* Remove TF_VAR style variables

tfenv is the new hotness, no more faffing with Terraform specific env
variables. See [1]

[1] https://github.com/cloudposse/tfenv

* Add ECS atlantis fargate task and atlantis.yml

* Add override vars for backwards compatibility

These vars are needed for backwards compat with currently provisioned
$stage.cloudposse.co resources, where we didn’t build the parent_domain
name via $STAGE.$NAMESPACE as we have with newer accounts e.g. evenco

* Bump ecs terraform-root-modules to 0.40.0

* Bump backing-services to root modules 0.40.0

* Remove atlantis triggers for changing resources

We will add these PR at a time when there isn’t a live demo going on.